+++ lang/syn/src/codegen/accounts/constraints.rs
@@ -142,54 +142,10 @@ pub fn generate_constraint_init(f: &Field, c: &ConstraintInitGroup) -> proc_macr
 
 pub fn generate_constraint_zeroed(f: &Field, _c: &ConstraintZeroed) -> proc_macro2::TokenStream {
     let field = &f.ident;
-    let (account_ty, account_wrapper_ty, _) = parse_ty(f);
-    let (account_combined_ty, try_from) = match &f.ty {
-        Ty::AccountInfo => (
-            quote! {
-                #account_wrapper_ty<#account_ty>
-            },
-            quote! { #field },
-        ),
-        Ty::Account(AccountTy { boxed, .. }) => {
-            if *boxed {
-                (
-                    quote! {
-                        Box<#account_wrapper_ty<#account_ty>>
-                    },
-                    quote! {
-                        Box::new(#account_wrapper_ty::try_from_unchecked(
-                            &#field,
-                        )?)
-                    },
-                )
-            } else {
-                (
-                    quote! {
-                        #account_wrapper_ty<#account_ty>
-                    },
-                    quote! {
-                        #account_wrapper_ty::try_from_unchecked(
-                            &#field,
-                        )?
-                    },
-                )
-            }
-        }
-        _ => (
-            quote! {
-                #account_wrapper_ty<#account_ty>
-            },
-            quote! {
-                #account_wrapper_ty::try_from_unchecked(
-                    program_id,
-                    &#field,
-                )?
-
-            },
-        ),
-    };
+    let ty_decl = f.ty_decl();
+    let from_account_info = f.from_account_info(None);
     quote! {
-        let #field: #account_combined_ty = {
+        let #field: #ty_decl = {
             let mut __data: &[u8] = &#field.try_borrow_data()?;
             let mut __disc_bytes = [0u8; 8];
             __disc_bytes.copy_from_slice(&__data[..8]);
@@ -197,7 +153,7 @@ pub fn generate_constraint_zeroed(f: &Field, _c: &ConstraintZeroed) -> proc_macr
             if __discriminator != 0 {
                 return Err(anchor_lang::__private::ErrorCode::ConstraintZero.into());
             }
-            #try_from
+            #from_account_info
         };
     }
 }
@@ -409,67 +365,6 @@ fn generate_constraint_seeds(f: &Field, c: &ConstraintSeedsGroup) -> proc_macro2
     }
 }
 
-fn parse_ty(f: &Field) -> (proc_macro2::TokenStream, proc_macro2::TokenStream, bool) {
-    match &f.ty {
-        Ty::ProgramAccount(ty) => {
-            let ident = &ty.account_type_path;
-            (
-                quote! {
-                    #ident
-                },
-                quote! {
-                    anchor_lang::ProgramAccount
-                },
-                false,
-            )
-        }
-        Ty::Account(ty) => {
-            let ident = &ty.account_type_path;
-            (
-                quote! {
-                    #ident
-                },
-                quote! {
-                    anchor_lang::Account
-                },
-                false,
-            )
-        }
-        Ty::Loader(ty) => {
-            let ident = &ty.account_type_path;
-            (
-                quote! {
-                    #ident
-                },
-                quote! {
-                    anchor_lang::Loader
-                },
-                true,
-            )
-        }
-        Ty::CpiAccount(ty) => {
-            let ident = &ty.account_type_path;
-            (
-                quote! {
-                    #ident
-                },
-                quote! {
-                    anchor_lang::CpiAccount
-                },
-                false,
-            )
-        }
-        Ty::AccountInfo => (
-            quote! {
-                AccountInfo
-            },
-            quote! {},
-            false,
-        ),
-        _ => panic!("Invalid type for initializing a program derived address"),
-    }
-}
-
 pub fn generate_init(
     f: &Field,
     seeds_with_nonce: proc_macro2::TokenStream,
@@ -478,69 +373,8 @@ pub fn generate_init(
     kind: &InitKind,
 ) -> proc_macro2::TokenStream {
     let field = &f.ident;
-    let (account_ty, account_wrapper_ty, is_zero_copy) = parse_ty(f);
-
-    let (combined_account_ty, try_from) = match f.ty {
-        Ty::AccountInfo => (
-            quote! {
-                AccountInfo
-            },
-            quote! {
-                #field.to_account_info()
-            },
-        ),
-        Ty::Account(AccountTy { boxed, .. }) => {
-            if boxed {
-                (
-                    quote! {
-                        Box<#account_wrapper_ty<#account_ty>>
-                    },
-                    quote! {
-                        Box::new(#account_wrapper_ty::try_from_unchecked(
-                            &#field.to_account_info(),
-                        )?)
-                    },
-                )
-            } else {
-                (
-                    quote! {
-                        #account_wrapper_ty<#account_ty>
-                    },
-                    quote! {
-                        #account_wrapper_ty::try_from_unchecked(
-                            &#field.to_account_info(),
-                        )?
-                    },
-                )
-            }
-        }
-        _ => {
-            let owner_addr = match &kind {
-                InitKind::Program { .. } => {
-                    quote! {
-                        program_id
-                    }
-                }
-                _ => {
-                    quote! {
-                        &anchor_spl::token::ID
-                    }
-                }
-            };
-            (
-                quote! {
-                    #account_wrapper_ty<#account_ty>
-                },
-                quote! {
-                    #account_wrapper_ty::try_from_unchecked(
-                        #owner_addr,
-                        &#field.to_account_info(),
-                    )?
-                },
-            )
-        }
-    };
-
+    let ty_decl = f.ty_decl();
+    let from_account_info = f.from_account_info(Some(kind));
     match kind {
         InitKind::Token { owner, mint } => {
             let create_account = generate_create_account(
@@ -550,7 +384,7 @@ pub fn generate_init(
                 seeds_with_nonce,
             );
             quote! {
-                let #field: #combined_account_ty = {
+                let #field: #ty_decl = {
                     // Define payer variable.
                     #payer
 
@@ -567,7 +401,7 @@ pub fn generate_init(
                     };
                     let cpi_ctx = CpiContext::new(cpi_program, accounts);
                     anchor_spl::token::initialize_account(cpi_ctx)?;
-                    let mut pa: #combined_account_ty = #try_from;
+                    let mut pa: #ty_decl = #from_account_info;
                     pa
                 };
             }
@@ -580,7 +414,7 @@ pub fn generate_init(
                 seeds_with_nonce,
             );
             quote! {
-                let #field: #combined_account_ty = {
+                let #field: #ty_decl = {
                     // Define payer variable.
                     #payer
 
@@ -595,7 +429,7 @@ pub fn generate_init(
                     };
                     let cpi_ctx = CpiContext::new(cpi_program, accounts);
                     anchor_spl::token::initialize_mint(cpi_ctx, #decimals, &#owner.to_account_info().key, None)?;
-                    let mut pa: #combined_account_ty = #try_from;
+                    let mut pa: #ty_decl = #from_account_info;
                     pa
                 };
             }
@@ -604,18 +438,21 @@ pub fn generate_init(
             let space = match space {
                 // If no explicit space param was given, serialize the type to bytes
                 // and take the length (with +8 for the discriminator.)
-                None => match is_zero_copy {
-                    false => {
-                        quote! {
-                            let space = 8 + #account_ty::default().try_to_vec().unwrap().len();
+                None => {
+                    let account_ty = f.account_ty();
+                    match matches!(f.ty, Ty::Loader(_)) {
+                        false => {
+                            quote! {
+                                let space = 8 + #account_ty::default().try_to_vec().unwrap().len();
+                            }
+                        }
+                        true => {
+                            quote! {
+                                let space = 8 + anchor_lang::__private::bytemuck::bytes_of(&#account_ty::default()).len();
+                            }
                         }
                     }
-                    true => {
-                        quote! {
-                            let space = 8 + anchor_lang::__private::bytemuck::bytes_of(&#account_ty::default()).len();
-                        }
-                    }
-                },
+                }
                 // Explicit account size given. Use it.
                 Some(s) => quote! {
                     let space = #s;
@@ -639,7 +476,7 @@ pub fn generate_init(
                     #space
                     #payer
                     #create_account
-                    let mut pa: #combined_account_ty = #try_from;
+                    let mut pa: #ty_decl = #from_account_info;
                     pa
                 };
             }